zoom_manager/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
tqdm==4.66.1
setuptools==75.3.0